wip: use willow-rs types
