+++ Cargo.toml
@@ -1,17 +1,17 @@
 [workspace]
 members = [
-  "iroh",
-  "iroh-blobs",
-  "iroh-base",
-  "iroh-dns-server",
-  "iroh-gossip",
-  "iroh-metrics",
-  "iroh-net",
-  "iroh-docs",
-  "iroh-test",
-  "iroh-net/bench",
-  "iroh-cli",
-  "iroh-willow",
+    "iroh",
+    "iroh-blobs",
+    "iroh-base",
+    "iroh-dns-server",
+    "iroh-gossip",
+    "iroh-metrics",
+    "iroh-net",
+    "iroh-docs",
+    "iroh-test",
+    "iroh-net/bench",
+    "iroh-cli",
+    "iroh-willow",
 ]
 resolver = "2"
 
@@ -43,9 +43,9 @@ missing_debug_implementations = "warn"
 unused-async = "warn"
 
 [patch.crates-io]
-# willow-data-model = { path = "../willow-rs/data-model" }
-# willow-encoding = { path = "../willow-rs/encoding" }
-# meadowcap = { path = "../willow-rs/meadowcap" }
-willow-data-model = { git = "https://github.com/Frando/willow-rs.git", branch = "iroh" }
-willow-encoding  = { git = "https://github.com/Frando/willow-rs.git", branch = "iroh" }
-meadowcap  = { git = "https://github.com/Frando/willow-rs.git", branch = "iroh" }
+willow-data-model = { path = "../willow-rs/data-model" }
+willow-encoding = { path = "../willow-rs/encoding" }
+meadowcap = { path = "../willow-rs/meadowcap" }
+# willow-data-model = { git = "https://github.com/Frando/willow-rs.git", branch = "iroh" }
+# willow-encoding  = { git = "https://github.com/Frando/willow-rs.git", branch = "iroh" }
+# meadowcap  = { git = "https://github.com/Frando/willow-rs.git", branch = "iroh" }
