+++ iroh-willow/src/proto.rs
@@ -1,8 +1,9 @@
 pub mod challenge;
+pub mod data_model;
 pub mod grouping;
 pub mod keys;
 pub mod meadowcap;
 pub mod pai;
-// pub mod sync;
-pub mod data_model;
+pub mod sync;
+pub mod wgps;
 pub use self::data_model as willow;
